formatCollapsed: collapse frames with different line numbers
formatCollapsed() was deduplicating frames only on the basis of their
internal index, which depends on the exact line number at which the
event occurred. So the output often contained apparent duplicates.

Instead, build an array of lines, with the text of each line in the key
of the array, and the count in the value. Finally, combine the lines.

Change-Id: I216fc6b20f0b34c0586e3f80fea904c2aa7d0b79