Mark truncated stack frames
Add a fake function excimer_truncated() as the ultimate parent of
truncated stack frames, so that the truncation will be apparent to the
end user, and so that truncated frames will be grouped together in a
flame graph.

The idea is for this to replace the post-processing done in
Wikimedia\MWConfig\Profiler::excimerFlushCallback().

Change-Id: I9caabcdffd1e9e6e53e8f911a40b0a575aeb3b0b