Implement partial macOS support
Excimer's functionality is backed by POSIX timers, which are not
available on macOS. One alternative way to reimplement this
functionality would be to use Apple's Grand Central Dispatch[1], but
this API is exclusive to Apple systems, which would complicate testing.

So, attempt to replicate the functionality by leveraging kqueue instead,
which is available not just on macOS, but also on BSDs. This should make
it easier to test the behavior even if one does not have access to an
Apple device. To make this possible, introduce a new os_timer
abstraction for timers backed by either a POSIX timer or kqueue, and use
it in excimer_timer.[2]

Sadly, it does not seem possible to make CPU profiling work with the
kqueue backend - but it should be usable for wall-time profiling.

[1] https://apple.github.io/swift-corelibs-libdispatch/tutorial/
[2] https://man.freebsd.org/cgi/man.cgi?query=kqueue&sektion=2

Bug: T316540
Change-Id: If0ec09f81138934abd378dec7e069aa5b7449a80