Fix staggering of the initial profile
The period is a floating-point number of seconds, often less than 1,
so when calculating the initial stagger time, it's not appropriate to
round it down to the nearest integer. php_mt_rand_range() returns an
integer. So, partially revert commit e5ad992064ec9abdbf7c8d.

The right way to select a number from a floating-point interval is with
php_random_gammasection_closed_open(), but we can switch to that once
PHP 8.3 is required, there's no urgency.

Add a regression test.

Bug: T378811
Change-Id: Ia9f10ff671e029f248b24887d044b47f3878b315