Fix memory leaks in excimer_log_get_speedscope_data
Confirmed the bug and verified the fix using valgrind with the supplied
test script.

zend_hash_add_new() can take a reference to str_key so we need to
release our reference, not destroy the string.

ht_indexes_by_key is purely internal so it can be safely destroyed.
zend_array_destroy() existed in PHP 7.1 and it frees both the contained
data and the HashTable.

Bug: T392788
Change-Id: Ia360a416363e567ea88dde8e62b2aa833332fdff